requirements.txt
# Core Streamlit app (minimal for startup)
streamlit==1.39.0
python-dotenv==1.0.1
openai==1.12.0
httpx>=0.24.0

# Database
psycopg2-binary==2.9.9
pgvector==0.3.6

# Text processing
pypdf==5.1.0

# Math libraries
numpy==1.26.4